Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
requests>=2.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
httpx>=0.28.0